% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_clusters.R
\name{align_clusters_matrix}
\alias{align_clusters_matrix}
\title{Align multiple clusterings in matrix format}
\usage{
align_clusters_matrix(
  clust_matrix,
  clust_ref,
  clust_names = NULL,
  handle_unmatched = "label",
  verbose = FALSE
)
}
\description{
Align multiple clusterings in matrix format
}
\keyword{internal}
