% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizSpatial.R
\name{basicPlotFormat}
\alias{basicPlotFormat}
\title{Plot formatting for presentation and publication}
\usage{
basicPlotFormat(
  p,
  theme_overall = "bw",
  base_size = 8,
  legend.position = "top",
  x.text.blank = F,
  y.text.blank = F,
  axis.title.bold = T,
  legend.title.bold = T,
  axis.title.x.blank = F,
  axis.title.y.blank = F,
  legend.title.blank = F,
  legend.spacing.x = 0,
  legend.key.spacing = 0,
  legend.key.spacing.y = 0,
  legend.box.margin = 0,
  legend.box.spacing = 0,
  legend.key.size = 8
)
}
\arguments{
\item{p}{A ggplot2 object.}

\item{theme_overall}{Theme of p.}

\item{base_size}{base_size of theme.}

\item{legend.position}{Legend position.}

\item{x.text.blank}{Logical. Set x axis text to blank.}

\item{y.text.blank}{Logical. Set y axis text to blank.}

\item{axis.title.bold}{Logical. Set title to be bold.}

\item{legend.title.bold}{Logical. Set legend title to be bold.}

\item{axis.title.x.blank}{Logical. Set x axis title to be bold.}

\item{axis.title.y.blank}{Logical. Set y axis title to be bold.}

\item{legend.title.blank}{Logical. Set legend title to be blank.}

\item{legend.spacing.x}{Legend spacing.}

\item{legend.key.spacing}{Legend key spacing.}

\item{legend.key.spacing.y}{Legend key spacing on vertical direction.}

\item{legend.box.margin}{Legend box margin.}

\item{legend.box.spacing}{Legend box spacing.,}

\item{legend.key.size}{Legend key size.}
}
\value{
A ggplot2 object.
}
\description{
Plot formatting for presentation and publication
}
