% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeCorMat.R
\name{cellTypeCorMat}
\alias{cellTypeCorMat}
\alias{cellTypeCorMat,matrix-method}
\alias{cellTypeCorMat,data.frame-method}
\alias{cellTypeCorMat,SingleCellExperiment-method}
\title{Create cell type correlation matrices}
\usage{
cellTypeCorMat(input, mask = NULL, reducedDimName = "PhiSpace", groupBy = NULL)

\S4method{cellTypeCorMat}{matrix}(input, mask = NULL, reducedDimName = "PhiSpace", groupBy = NULL)

\S4method{cellTypeCorMat}{data.frame}(input, mask = NULL, reducedDimName = "PhiSpace", groupBy = NULL)

\S4method{cellTypeCorMat}{SingleCellExperiment}(input, mask = NULL, reducedDimName = "PhiSpace", groupBy = NULL)
}
\arguments{
\item{input}{A \code{matrix} containing numeric data for correlation computation.}

\item{mask}{A zero-one matrix with the same number of rows as the input. Optional. If provided,
each column of \code{mask} defines a subset of rows used to compute a separate correlation matrix.}

\item{reducedDimName}{A character string specifying the name of the reduced dimension to use for \code{SingleCellExperiment} objects. Defaults to \code{"PhiSpace"}.}

\item{groupBy}{A column of \code{colData(input)} when input is an SCE object. If provided, mask will be defined according to \code{colData(input)[,groupBy]}.}
}
\value{
If \code{mask} is provided, returns a list of correlation matrices (one per column of \code{mask}).
Otherwise, returns a single correlation matrix.
}
\description{
A generic function to compute a correlation matrix for different input types, including \code{matrix},
\code{data.frame} and \code{SingleCellExperiment} objects.
}
\details{
This method calculates the correlation matrix directly for the input \code{matrix}.

This method first converts the \code{data.frame} to a \code{matrix} and then calculates the correlation matrix.

This method uses the reduced dimension \code{"PhiSpace"} of the \code{SingleCellExperiment} object to compute the correlation matrix.
If \code{"PhiSpace"} is not available, an error is raised.
}
