% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLRnorm.R
\name{CLRnorm}
\alias{CLRnorm}
\title{Center log ratio normalisation}
\usage{
CLRnorm(sce, assayName = "counts", targetAssay = "data", offset = 1)
}
\arguments{
\item{sce}{SCE object}

\item{assayName}{Assay to normalise}

\item{targetAssay}{Name of normalised assay}

\item{offset}{Offset value added to raw counts}
}
\value{
SCE object
}
\description{
Center log ratio normalisation
}
