% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeY.R
\name{codeY}
\alias{codeY}
\title{Convert discrete annotations to dummy variable matrices.}
\usage{
codeY(sce, phenotypes, method = c("-1,1", "0,1"))
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{phenotypes}{Which phenotype annotations (e.g. cell types) to convert.}

\item{method}{Which coding to use, either (-1,1) or (0,1).}
}
\value{
\item{YY}{A dummy variable matrix encoding discrete annotation}
}
\description{
Convert discrete annotations to dummy variable matrices.
}
