% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeY.R
\name{codeY_vec}
\alias{codeY_vec}
\title{Convert discrete annotations to dummy variable matrices.}
\usage{
codeY_vec(vec, rowNames = NULL, method = c("-1,1", "0,1"))
}
\arguments{
\item{vec}{Factor contanining class labels.}

\item{rowNames}{Optional. Row names for output dummy matrix (eg cell barcode).}

\item{method}{Which coding to use, either (-1,1) or (0,1).}
}
\value{
\item{YY}{A dummy variable matrix encoding discrete annotation}
}
\description{
Convert discrete annotations to dummy variable matrices.
}
