% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUMAP.R
\name{computeUMAP}
\alias{computeUMAP}
\title{Compute UMAP.}
\usage{
computeUMAP(
  dat,
  computPC = TRUE,
  ncomp = 30,
  center = TRUE,
  scale = FALSE,
  config = NULL
)
}
\arguments{
\item{dat}{Input data.}

\item{computPC}{Whether to compute PC first.}

\item{ncomp}{Number of PC components}

\item{center}{Center or not.}

\item{scale}{Scale or not.}

\item{config}{Configuration of UMAP (see \code{umap} package).}
}
\value{
UMAP coordinates.
}
\description{
Compute UMAP.
}
