% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP.R
\name{computUMAP}
\alias{computUMAP}
\title{Compute UMAP.}
\usage{
computUMAP(dat, computPC = TRUE, ncomp = 30, center = TRUE, scale = FALSE)
}
\arguments{
\item{dat}{Input data.}

\item{computPC}{Whether to compute PC first.}

\item{ncomp}{Number of PC components}

\item{center}{Center or not.}

\item{scale}{Scale or not.}

\item{labs}{Labels.}
}
\value{
UMAP coordinates.
}
\description{
Compute UMAP.
}
