% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVTune_ncomp.R
\name{CVTune_ncomp}
\alias{CVTune_ncomp}
\title{Tune ncomp using cross-validation.}
\usage{
CVTune_ncomp(
  reference,
  assayName = "logcounts",
  phenotypes = NULL,
  YY = NULL,
  center = TRUE,
  scale = FALSE,
  ncompLimits = c(1, 100),
  gridSize = 20,
  ncompGrid = NULL,
  selectedFeat = NULL,
  Kfolds = 5,
  regMethod = "PCA",
  labelCode = "-1,1",
  Ncores = 1,
  seed = 5202056
)
}
\arguments{
\item{reference}{\code{SingleCellExperiment} object.}

\item{assayName}{Character.}

\item{phenotypes}{Character.}

\item{YY}{Matrix.}

\item{center}{Logic.}

\item{scale}{Logic.}

\item{ncompLimits}{Vector.}

\item{gridSize}{Integer.}

\item{ncompGrid}{Vector.}

\item{selectedFeat}{Vector.}

\item{Kfolds}{Integer.}

\item{regMethod}{Character.}

\item{labelCode}{Character.}

\item{Ncores}{Integer.}

\item{seed}{Integer.}
}
\value{
A list.
}
\description{
Tune ncomp using cross-validation.
}
