% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVTune_nfeat.R
\name{CVTune_nfeat}
\alias{CVTune_nfeat}
\title{Tune nfeat using cross-validation.}
\usage{
CVTune_nfeat(
  reference,
  assayName = "logcounts",
  phenotypes = NULL,
  YY = NULL,
  ncomp = NULL,
  impScores = NULL,
  center = TRUE,
  scale = FALSE,
  nfeatLimits = c(10, 15000),
  gridSize = 30,
  nfeatV = NULL,
  Kfolds = 5,
  regMethod = "PCA",
  labelCode = "-1,1",
  seed = 5202056,
  Ncores = 1
)
}
\arguments{
\item{reference}{\code{SingleCellExperiment} object.}

\item{assayName}{Character.}

\item{phenotypes}{Vector.}

\item{YY}{Matrix.}

\item{ncomp}{Integer.}

\item{impScores}{Matrix.}

\item{center}{Logic.}

\item{scale}{Logic.}

\item{nfeatLimits}{Vector.}

\item{gridSize}{Integer.}

\item{nfeatV}{Vector.}

\item{Kfolds}{Integer.}

\item{regMethod}{Character.}

\item{labelCode}{Character.}

\item{seed}{Integer.}

\item{Ncores}{Integer.}
}
\value{
A list.
}
\description{
Tune nfeat using cross-validation.
}
