% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVTuneNcomp.R
\name{CVTuneNcomp}
\alias{CVTuneNcomp}
\title{Tune ncomp using cross-validation.}
\usage{
CVTuneNcomp(
  reference,
  YY = NULL,
  labPerSample = NULL,
  labelName = NULL,
  ncompLimits = c(1, 100),
  gridSize = 20,
  ncompGrid = NULL,
  selectFeat = NULL,
  assayName = "logcounts",
  Kfolds = 5,
  seed = 5202056,
  regMethod = "PCA",
  mode = "supervised",
  normYY = F,
  Ncores = 2,
  labelCode = "-1,1"
)
}
\arguments{
\item{labelCode}{}
}
\description{
Tune ncomp using cross-validation.
}
