% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVTuneNselect.R
\name{CVTuneNselect}
\alias{CVTuneNselect}
\title{Tune Nselect using cross-validation.}
\usage{
CVTuneNselect(
  reference,
  ncomp,
  impScores,
  YY = NULL,
  labPerSample = NULL,
  labelName = NULL,
  assayName = "logcounts",
  NselectLimits = c(10, 15000),
  gridSize = 30,
  NselectV = NULL,
  Kfolds = 5,
  seed = 5202056,
  regMethod = "PCA",
  mode = "supervised",
  Ncores = 5,
  normYY = F,
  labelCode = "-1,1"
)
}
\arguments{
\item{labelCode}{}
}
\description{
Tune Nselect using cross-validation.
}
