% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankFeatures.R
\name{.fit_pls}
\alias{.fit_pls}
\title{Internal function to fit PLS or PLSDA}
\usage{
.fit_pls(XX, YY, ncomp, center, scale, method)
}
\description{
Internal function to fit PLS or PLSDA
}
\keyword{internal}
