% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPhiSpace.R
\name{.select_optimal_k}
\alias{.select_optimal_k}
\title{Internal function to select optimal k}
\usage{
.select_optimal_k(
  pc_scores,
  k_range,
  method,
  nstart,
  iter.max,
  algorithm,
  silhouette_max_cells = Inf
)
}
\description{
Internal function to select optimal k
}
\keyword{internal}
