% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doubleCent}
\alias{doubleCent}
\title{Double center a matrix by column and row means, resulting in a new matrix with zero row and column means.}
\usage{
doubleCent(X)
}
\arguments{
\item{X}{A matrix or an object convertible to one.}
}
\value{
A double-centred matrix.
}
\description{
Double center a matrix by column and row means, resulting in a new matrix with zero row and column means.
}
