% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErr.R
\name{getErr}
\alias{getErr}
\title{Compute different errors for cross-validation.}
\usage{
getErr(
  ncompGrid,
  XX,
  YY,
  XXtest,
  YYtest,
  labPerSample = NULL,
  selectFeat = NULL,
  regMethod = "PLS",
  assayName = "rank",
  mode = "supervised",
  normYY = F
)
}
\arguments{
\item{normYY}{}
}
\description{
Compute different errors for cross-validation.
}
