% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErr_nfeat.R
\name{getErr_nfeat}
\alias{getErr_nfeat}
\title{Calculate errors for the cross-validation for selecting nfeat.}
\usage{
getErr_nfeat(
  nfeatV,
  impScores,
  ncomp,
  XXtrain,
  YYtrain,
  XXtest,
  YYtest,
  center = TRUE,
  scale = FALSE,
  regMethod = "PLS",
  assayName = "logcounts"
)
}
\arguments{
\item{nfeatV}{Vector.}

\item{impScores}{Matrix.}

\item{ncomp}{Integer.}

\item{XXtrain}{Matrix.}

\item{YYtrain}{Matrix.}

\item{XXtest}{Matrix.}

\item{YYtest}{Matrix.}

\item{center}{Logic.}

\item{scale}{Logic.}

\item{regMethod}{Character.}

\item{assayName}{Character.}
}
\value{
A list.
}
\description{
Calculate errors for the cross-validation for selecting nfeat.
}
