% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrNselect.R
\name{getErrNselect}
\alias{getErrNselect}
\title{Calculate errors for the cross-validation for selecting Nselect.}
\usage{
getErrNselect(
  NselectV,
  orderByCol,
  ncomp,
  XX,
  YY,
  XXtest,
  YYtest,
  labPerSample = NULL,
  regMethod = "PLS",
  assayName = "logcounts",
  mode = "supervised",
  normYY = F
)
}
\arguments{
\item{normYY}{}
}
\description{
Calculate errors for the cross-validation for selecting Nselect.
}
