% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeepCommonGenes.R
\name{KeepCommonGenes}
\alias{KeepCommonGenes}
\title{Keep common genes of two SingleCellExperiment objects.}
\usage{
KeepCommonGenes(reference, query, dots2hyphen = F)
}
\arguments{
\item{reference}{SingleCellExperiment object.}

\item{query}{SingleCellExperiment object.}

\item{dots2hyphen}{Logic.}
}
\value{
A list containing updated SingleCellExperiment obejcts with matching features.
}
\description{
Keep common genes of two SingleCellExperiment objects.
}
