% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Soft phenotyping of query assay.}
\usage{
phenotype(
  phenoAssay,
  atlas_re,
  assayName = "rank",
  scaleMethod = c("byQuery", "byRef"),
  scale = FALSE
)
}
\arguments{
\item{phenoAssay}{Matrix. Query assay to phenotype.}

\item{atlas_re}{List.}

\item{assayName}{Character.}

\item{scaleMethod}{Character.}

\item{scale}{Logic.}
}
\value{
A list containing
\item{Yhat}{}
\item{Proj}{}
}
\description{
Soft phenotyping of query assay.
}
