% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPhiSpace.R
\name{plot.PhiSpaceClustering}
\alias{plot.PhiSpaceClustering}
\title{Plot method for PhiSpaceClustering objects}
\usage{
\method{plot}{PhiSpaceClustering}(x, type = c("pca", "elbow", "silhouette", "variance"), ...)
}
\arguments{
\item{x}{A PhiSpaceClustering object}

\item{type}{Character string specifying plot type. Options: "pca" (PC1 vs PC2),
"elbow" (elbow plot if k_range was used), "silhouette" (silhouette plot if
k_range was used), "variance" (PCA variance explained). Default is "pca".}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for PhiSpaceClustering objects
}
