% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhiSpaceHeatMap.R
\name{plotPhiSpaceHeatMap}
\alias{plotPhiSpaceHeatMap}
\title{Plot PhiSpace scores as a heatmap}
\usage{
plotPhiSpaceHeatMap(
  PhiSpaceScore,
  phenoDict = NULL,
  phenotypes = NULL,
  queryLabs = NULL,
  refLvls = NULL,
  queryLvls = NULL,
  cluster_columns = FALSE,
  cluster_rows = FALSE,
  show_row_names = FALSE,
  show_column_names = TRUE,
  row_title_rot = 0,
  ...
)
}
\arguments{
\item{PhiSpaceScore}{Matrix.}

\item{phenoDict}{}

\item{phenotypes}{Optional. Types of phenotypes. If provided, has to be of the same length as phenoDict.}

\item{queryLabs}{Character vector.}

\item{refLvls}{Character vector.}

\item{queryLvls}{Character vector.}

\item{cluster_columns}{Logic.}

\item{cluster_rows}{Logic.}

\item{show_row_names}{Logic.}

\item{show_column_names}{Logic.}

\item{row_title_rot}{Rotation (angle) of row titles, inherited from \code{ComplexHeatmap::Heatmap}}

\item{...}{Additional parameters adjusting the \code{ComplexHeatmap} object}
}
\value{
A heatmap of PhiSpace scores.
}
\description{
Plot PhiSpace scores as a heatmap
}
