% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSankey.R
\name{plotSankey}
\alias{plotSankey}
\title{Plot Sankey diagram for two classification results.}
\usage{
plotSankey(class1, class2, fontsize = 12)
}
\arguments{
\item{fontsize}{}
}
\description{
Plot Sankey diagram for two classification results.
}
