% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSankey3.R
\name{plotSankey3}
\alias{plotSankey3}
\title{Plot Sankey diagram for three hard classification results.}
\usage{
plotSankey3(class1, class2, class3, add = TRUE, fontsize = 12)
}
\arguments{
\item{class1}{Vector.}

\item{class2}{Vector.}

\item{class3}{Vector.}

\item{add}{Logic.}

\item{fontsize}{Numeric.}
}
\description{
Plot Sankey diagram for three hard classification results.
}
