% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.fit.R
\name{pls.fit}
\alias{pls.fit}
\title{Title}
\usage{
pls.fit(X, Y, ncomp)
}
\arguments{
\item{X}{Matrix. Scaled predictor matrix}

\item{Y}{Matrix. Scaled response matrix}

\item{ncomp}{Integer.}

\item{center}{Logic.}
}
\value{
A list containing
\item{coefficients}{Regression coefficient matrices.}
}
\description{
Same as \code{pls::kernelpls.fit}, but replacing a full eigen decomposition by a partial one via \code{rARPACK::eigs_sym}.
}
