% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulk.R
\name{pseudoBulk}
\alias{pseudoBulk}
\title{Converting single cells to pseudo-bulks.}
\usage{
pseudoBulk(
  sce,
  phenotypes = NULL,
  response = NULL,
  clusterid = NULL,
  assayName = "counts",
  resampSizes = 100,
  proportion = NULL,
  nPool = 15,
  seed = 904800
)
}
\arguments{
\item{sce}{\code{SingleCellExpeirment} object.}

\item{phenotypes}{Phenotypes to predict. Can be multiple.}

\item{response}{Response matrix. Can be continuous.}

\item{clusterid}{Column name of colData for specifying groups for pseudobulking.}

\item{assayName}{Assay used for pseudobulking.}

\item{resampSizes}{How many pseudo-bulk resamples to generate from each cluster.}

\item{proportion}{Used to determine unequal resampSizes.}

\item{nPool}{How many cells to use for pseudobulking.}

\item{seed}{Random seeds}
}
\value{
An updated \code{SingleCellExpeirment} object with a pseudobulk assay.
}
\description{
Converting single cells to pseudo-bulks.
}
