% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankTransf.R
\name{RankTransf}
\alias{RankTransf}
\title{Rank transform of a \code{SingleCellExperiment} object.}
\usage{
RankTransf(sce, assayname = "counts", targetAssay = "rank", sparse = TRUE)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object.}

\item{assayname}{Character.}

\item{targetAssay}{Character.}

\item{sparse}{Logic.}
}
\value{
An updated \code{SingleCellExperiment} object with a rank transformed assay.
}
\description{
Rank transform of a \code{SingleCellExperiment} object.
}
