% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcTransf.R
\name{rcTransf}
\alias{rcTransf}
\title{Relative counts normalisation.}
\usage{
rcTransf(sce, assayName = "counts", targetAssay = "data")
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{assayName}{Character.}

\item{targetAssay}{Character.}
}
\value{
A SingleCellExperiment object containing a new assay after relative counts normalisation.
}
\description{
Relative counts normalisation.
}
