% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reNameCols}
\alias{reNameCols}
\title{Turn a matrix to a data.frame withe newly defined colnames.}
\usage{
reNameCols(mat, key = "comp")
}
\arguments{
\item{mat}{Matrix.}

\item{key}{String.}
}
\value{
Data frame.
}
\description{
Turn a matrix to a data.frame withe newly defined colnames.
}
