% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSampler.R
\name{sample_grid}
\alias{sample_grid}
\title{Grid-based spatial sampling}
\usage{
sample_grid(
  coordinates,
  n_target,
  grid_size,
  min_cells_per_region,
  balance_regions,
  verbose = TRUE
)
}
\arguments{
\item{coordinates}{Spatial coordinates}

\item{n_target}{Number of neighbours}

\item{grid_size}{Grid size}

\item{min_cells_per_region}{Mimimum cells per region}

\item{balance_regions}{Balance regions}

\item{verbose}{Whether to report progress with \code{\link[=message]{message()}}}
}
\description{
Grid-based spatial sampling
}
