% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSampler.R
\name{sample_random}
\alias{sample_random}
\title{Random sampling (baseline method)}
\usage{
sample_random(coordinates, n_target)
}
\arguments{
\item{coordinates}{Spatial coordiantes}

\item{n_target}{Number of targets}
}
\description{
Random sampling (baseline method)
}
