% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scranTransf.R
\name{scranTransf}
\alias{scranTransf}
\title{Run scran normalisation}
\usage{
scranTransf(sce, assayName = "counts", clusts = NULL)
}
\arguments{
\item{sce}{SCE object.}

\item{assayName}{Assay name.}

\item{clusts}{Cluster labels (optional).}
}
\value{
SCE object.
}
\description{
Run scran normalisation
}
