% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeat.R
\name{selectFeat}
\alias{selectFeat}
\title{Select nfeat top features from each column (label) of impScores}
\usage{
selectFeat(impScores, nfeat)
}
\arguments{
\item{impScores}{Matrix. Each column corresponds to a phenotype (eg a cell type) and contains importance scores of features for predicting that phenotype.}

\item{nfeat}{Number of top features to select from each column of impScores.}
}
\value{
Vector of selected features.
}
\description{
Select nfeat top features from each column (label) of impScores
}
