% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split2}
\alias{split2}
\title{Create partitions of data for cross-validation.}
\usage{
split2(x, n)
}
\arguments{
\item{x}{Vector.}

\item{n}{Integer.}
}
\value{
A partition of index vector \code{x} to \code{n} folds.
}
\description{
Create partitions of data for cross-validation.
}
