% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsample from a sce object}
\usage{
subsample(sce, key = NULL, proportion = 0.1, minCellNum = 50, seed = 5202056)
}
\arguments{
\item{sce}{SCE object to be sampled from}

\item{key}{Required for stratefied subsampling}

\item{proportion}{Proportion of cells to sample}

\item{minCellNum}{Minimum number of cells in each category}

\item{seed}{Random seed}
}
\value{
A downsized SCE object
}
\description{
Subsample from a sce object
}
