% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPhiSpace.R
\name{summary.PhiSpaceClustering}
\alias{summary.PhiSpaceClustering}
\title{Summary method for PhiSpaceClustering objects}
\usage{
\method{summary}{PhiSpaceClustering}(object, ...)
}
\arguments{
\item{object}{A PhiSpaceClustering object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for PhiSpaceClustering objects
}
