% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC.R
\name{SuperPC}
\alias{SuperPC}
\title{Compute regression model using selected features.}
\usage{
SuperPC(
  reference,
  YY,
  ncomp,
  selectedFeat = NULL,
  assayName = "logcounts",
  regMethod = c("PCA", "PLS"),
  center = TRUE,
  scale = FALSE,
  sparse = TRUE,
  DRinfo = FALSE
)
}
\arguments{
\item{reference}{SingleCellExperiment object (the reference), or a cell by
gene matrix of the \code{assayName} assay.}

\item{YY}{Resonse matrix}

\item{ncomp}{Integer. Number of components}

\item{selectedFeat}{Character. Vector containing selected feature names}

\item{assayName}{Which assay in reference to use for prediction}

\item{regMethod}{Character. Regression method to use, either PCA or PLS.}

\item{center}{Logic.}

\item{scale}{Logic. Scale the predictor matrix or not}

\item{sparse}{Use sparse matrices or not.}

\item{DRinfo}{Logic. Whether to return dimension reduction information from PCA or PLS. Disable to save memory.}
}
\value{
A list containing regression input and output.
}
\description{
Compute regression model using selected features.
}
