% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdspc.fit.R
\name{svdspc.fit}
\alias{svdspc.fit}
\title{Principle component regression}
\usage{
svdspc.fit(X, Y, ncomp, center = TRUE, stripped = FALSE)
}
\arguments{
\item{stripped}{}
}
\description{
Same as \code{pls::svdpc.fit}, except using \code{rARPACK::svds} instead of \code{La.svd}.
}
