% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateLabel.R
\name{translateLabel}
\alias{translateLabel}
\title{Translate cell type labels from finer to coarser according to a dictionary.}
\usage{
translateLabel(text, dictionary, dFrom = "label.fine", dTo = "label.main")
}
\arguments{
\item{dTo}{}
}
\value{
A vector containing translated labels.
}
\description{
Translate cell type labels from finer to coarser according to a dictionary.
}
