% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunePhiSpace.R
\name{tunePhiSpace}
\alias{tunePhiSpace}
\title{Title}
\usage{
tunePhiSpace(
  reference,
  assayName = "logcounts",
  phenotypes = NULL,
  YY = NULL,
  ncomp = NULL,
  impScores = NULL,
  center = TRUE,
  scale = FALSE,
  tune_ncomp = TRUE,
  tune_nfeat = TRUE,
  ncompLimits = c(1, 100),
  ncompGridSize = 20,
  ncompGrid = NULL,
  selectedFeat = NULL,
  nfeatLimits = c(10, 15000),
  nfeatGridSize = 30,
  nfeatV = NULL,
  Kfolds = 5,
  regMethod = "PLS",
  labelCode = "-1,1",
  Ncores = 1,
  seed = 5202056
)
}
\arguments{
\item{reference}{\code{SingleCellExperiment} object.}

\item{assayName}{Character.}

\item{phenotypes}{Character.}

\item{YY}{Matrix.}

\item{ncomp}{Integer.}

\item{impScores}{Matrix.}

\item{center}{Logical.}

\item{scale}{Logical.}

\item{tune_ncomp}{Logic.}

\item{tune_nfeat}{Logic.}

\item{ncompLimits}{Vector.}

\item{ncompGridSize}{Integer.}

\item{ncompGrid}{Vector.}

\item{selectedFeat}{Vector.}

\item{nfeatLimits}{Vector.}

\item{nfeatGridSize}{Integer.}

\item{nfeatV}{Vector.}

\item{Kfolds}{Integer.}

\item{regMethod}{Character.}

\item{labelCode}{Character.}

\item{Ncores}{Integer.}

\item{seed}{Integer.}
}
\value{
A list
}
\description{
Title
}
