% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizSpatial.R
\name{VizSpatial}
\alias{VizSpatial}
\title{Visualising features (eg gene expression), metadata (eg cell type) or reducedDim component (eg PhiSpace cell type score) in 2D tissue space.}
\usage{
VizSpatial(
  obj,
  x_coord = "x",
  y_coord = "y",
  ptSize = 2,
  ptShape = 16,
  groupBy = NULL,
  feature = NULL,
  assay2use = NULL,
  reducedDim = NULL,
  reducedDim2use = "PhiSpace",
  censor = FALSE,
  quant = 0.5,
  legend.position = "right",
  legend.symb.size = NULL,
  legend.title = NULL,
  fsize = 14,
  reOrder = F,
  ...
)
}
\arguments{
\item{obj}{Input object; either \code{SingleCellExperiemnt} or \code{SpatialExperiment}. If \code{SingleCellExperiemnt}, then \code{colData(obj)} has to contain columns named by \code{x_coord} and \code{y_coord} below.}

\item{x_coord}{x coordinate name of a column of colData (if \code{type(ojb) == SingleCellExperiemnt}).}

\item{y_coord}{y coordinate name of a column of colData (if \code{type(ojb) == SingleCellExperiemnt}).}

\item{ptSize}{Size of point, representing spatial cell like objects (segmented cells, spots etc).}

\item{ptShape}{Shape of point.}

\item{groupBy}{Name of a metadata column to colour points by.}

\item{feature}{Name of a feature (eg gene) to colour points by.}

\item{assay2use}{Which assay to use when feature is specified.}

\item{reducedDim}{Name of reducedDim column to colour points by.}

\item{reducedDim2use}{Name of reducedDim layer to use when reducedDim is specified.}

\item{censor}{Logical. Whether to set smaller colour values to zero. Default is \code{FALSE}. (Recommend to set to \code{TRUE} when visualise PhiSpace scores,)}

\item{quant}{A value between 0 and 1. Colour values below \code{quant} will be set to zero if \code{censor=TRUE}.}

\item{legend.position}{Position of legened, one of "right", "bottome", "left" and "top".}

\item{legend.symb.size}{Legend symbol size (for discrete legend).}

\item{legend.title}{Alternative legend name to replace the original.}

\item{fsize}{Base font size of figure.}

\item{reOrder}{Logical. Whether to reorder points according to their values (ascending) or not. Set to be \code{TRUE} to avoid overplotting.}

\item{...}{Other arguments to pass to \code{basicPlotFormat}.}
}
\value{
A ggplot2 object.
}
\description{
The 2D spatial coordinates (x and y) should be stored as two columns in colData.
}
