% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroFeatQC.R
\name{zeroFeatQC}
\alias{zeroFeatQC}
\title{Minimal quality control (QC): remove features with all zero values.}
\usage{
zeroFeatQC(sce, assayName = "counts", verbose = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{assayName}{Character.}

\item{verbose}{Logical. Whether to report the result with \code{\link[=message]{message()}}
(default: TRUE).}
}
\value{
An updated SingleCellExperiment object after minimal QC.
}
\description{
Minimal quality control (QC): remove features with all zero values.
}
