% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_clusters.R
\name{align_clusters}
\alias{align_clusters}
\title{Align clustering results.}
\usage{
align_clusters(clust, clust_ref)
}
\arguments{
\item{clust}{Query clustering result, can be a numerical, character or factor vector.}

\item{clust_ref}{Reference clustering result.}
}
\value{
Re-aligned clustering result, represented as a factor vector.
}
\description{
Given a reference clustering result, represented by a vector of cluster labels, align a query clustering result to the reference one.
}
