% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classErr.R, R/utils.R
\name{classErr}
\alias{classErr}
\title{Calculate classification errors.}
\usage{
classErr(classQuery, classOriginal, labPerSample = NULL)

classErr(classQuery, classOriginal, labPerSample = NULL)
}
\arguments{
\item{classQuery}{Character vector.}

\item{classOriginal}{Character vector.}

\item{labPerSample}{Integer.}
}
\value{
A list.

A list containing:
\item{err}{}
\item{byClassErrs}{}
}
\description{
Calculate classification errors.

Calcualte classification errors.
}
