% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErr.R
\name{getErr}
\alias{getErr}
\title{Compute different errors for cross-validation.}
\usage{
getErr(
  ncompGrid,
  XXtrain,
  YYtrain,
  XXtest,
  YYtest,
  center = TRUE,
  scale = FALSE,
  selectedFeat = NULL,
  regMethod = "PLS",
  assayName = "rank",
  loss = c("Frobenius", "exponential")
)
}
\arguments{
\item{ncompGrid}{Vector.}

\item{XXtrain}{Matrix.}

\item{YYtrain}{Matrix.}

\item{XXtest}{Matrix.}

\item{YYtest}{Matrix.}

\item{center}{Logic.}

\item{scale}{Logic.}

\item{selectedFeat}{Vector.}

\item{regMethod}{Character.}

\item{assayName}{Character.}

\item{loss}{Type of loss function.}
}
\value{
A list
}
\description{
Compute different errors for cross-validation.
}
