% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErr.R
\name{getErr}
\alias{getErr}
\title{Compute different errors for cross-validation.}
\usage{
getErr(
  ncompGrid,
  XX,
  YY,
  XXtest,
  YYtest,
  labPerSample = NULL,
  selectedFeat = NULL,
  regMethod = "PLS",
  assayName = "rank",
  mode = "supervised",
  normYY = F
)
}
\arguments{
\item{ncompGrid}{Vector.}

\item{XX}{Matrix.}

\item{YY}{Matrix.}

\item{XXtest}{Matrix.}

\item{YYtest}{Matrix.}

\item{labPerSample}{Integer.}

\item{selectedFeat}{Vector.}

\item{regMethod}{Character.}

\item{assayName}{Character.}

\item{mode}{Character.}

\item{normYY}{Logic.}
}
\value{
A list
}
\description{
Compute different errors for cross-validation.
}
