% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPC.R
\name{getPC}
\alias{getPC}
\title{Principal component analysis (PCA) based on partial singular value decomposition (SVD).}
\usage{
getPC(X, ncomp, center = TRUE, scale = FALSE)
}
\arguments{
\item{scale}{}
}
\description{
Principal component analysis (PCA) based on partial singular value decomposition (SVD).
}
