% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPC.R
\name{getPC}
\alias{getPC}
\title{Principal component analysis (PCA) based on partial singular value decomposition (SVD).}
\usage{
getPC(X, ncomp, center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{Matrix.}

\item{ncomp}{Integer.}

\item{center}{Logic.}

\item{scale}{Logic.}
}
\value{
A list containing
\item{scores}{Score matrix for X.}
\item{loadings}{Laoding matrix for X.}
\item{sdev}{}
\item{totVar}{}
\item{props}{}
\item{accuProps}{}
\item{ncomp}{}
\item{selectFeat}{}
\item{reg_re}{}
}
\description{
Principal component analysis (PCA) based on partial singular value decomposition (SVD).
}
