% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepCommonGenes.R
\name{keepCommonGenes}
\alias{keepCommonGenes}
\title{Keep common genes of two SingleCellExperiment objects.}
\usage{
keepCommonGenes(reference, query, dots2hyphen = F)
}
\arguments{
\item{dots2hyphen}{}
}
\description{
Keep common genes of two SingleCellExperiment objects.
}
