% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTransf.R
\name{logTransf}
\alias{logTransf}
\title{Relative counts normalisation.}
\usage{
logTransf(
  sce,
  scalFactor = 10000,
  assayName = "counts",
  targetAssay = "data",
  use_log1p = FALSE,
  sparse = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{scalFactor}{Scale factor to inflate relative counts.}

\item{assayName}{Character.}

\item{targetAssay}{Character.}

\item{use_log1p}{Logical. Use shifted log (log1p) transform, using raw counts instead of relative counts.}

\item{sparse}{Store the transformed assay as sparse matrix or not.}
}
\value{
A SingleCellExperiment object containing a new assay after relative counts normalisation.
}
\description{
Relative counts normalisation.
}
