% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{matrixPlot}
\alias{matrixPlot}
\title{Score matrix plot}
\usage{
matrixPlot(
  scores,
  max_ncomp = NULL,
  comp_idx = NULL,
  colBy = NULL,
  pointAlpha = NULL,
  pointSize = 1,
  manualCol = NULL,
  manualAlpha = NULL,
  fsize = 14
)
}
\arguments{
\item{scores}{Matrix of scores to be plotted}

\item{max_ncomp}{Default NULL. Number of first components to plot. If specified, will override comp_idx.}

\item{comp_idx}{Default NULL. Indices of components to plot.}

\item{colBy}{Numeric or charactor vectors to specify colour of points.}

\item{pointAlpha}{Alpha value.}

\item{pointSize}{Point size.}

\item{manualCol}{Manual specification of colours.}

\item{manualAlpha}{Manual specification of alpha colours.}

\item{fsize}{Figure font size.}
}
\description{
Density and pairwise scatter plots for visualising score matrices.
}
