% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvr.R
\name{mvr}
\alias{mvr}
\title{Multivariate regression via principal component analysis (PCA) or partial least squares (PLS).}
\usage{
mvr(X, Y, ncomp, method = c("PCA", "PLS"), center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{Matrix.}

\item{Y}{Matrix.}

\item{ncomp}{Integer.}

\item{method}{Character.}

\item{center}{Logic.}

\item{scale}{Logic.}
}
\value{
A list containing
\item{coefficients}{Regression coefficient matrices.}
\item{Xmeans}{}
\item{Ymeans}{}
\item{ncomp}{}
\item{method}{}
}
\description{
Simplified version of \code{pls::mvr}, using computationally faster versions of PCA and PLS.
}
