% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvr.R
\name{mvr}
\alias{mvr}
\title{Multivariate regression via principal component analysis (PCA) or partial least squares (PLS).}
\usage{
mvr(X, Y, ncomp, method = c("PCA", "PLS"), center = T, scale = F)
}
\arguments{
\item{scale}{}
}
\description{
Simplified version of \code{pls::mvr}, using computationally faster versions of PCA and PLS.
}
