% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Soft phenotyping of query assay.}
\usage{
phenotype(assayToProj, atlas_re, assayName = "rank", scale = FALSE)
}
\arguments{
\item{scale}{}
}
\value{
A list containing
\item{Yhat}{}
\item{Proj}{}
}
\description{
Soft phenotyping of query assay.
}
