% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhiSpace-package.R
\docType{package}
\name{PhiSpace-package}
\alias{PhiSpace-package}
\title{PhiSpace: Phenotype space analysis of single-cell multiomics and spatial transcriptomics data}
\description{
Continous phenotyping of single-cell multiomics data using partial least squares regression.
}
\author{
\strong{Maintainer}: Jiadong Mao \email{chiatungmao@gmail.com} (\href{https://orcid.org/0000-0002-3818-1981}{ORCID})

}
\keyword{internal}
