% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhiSpaceR.R
\name{PhiSpace}
\alias{PhiSpace}
\title{PhiSpace continuous phenotyping}
\usage{
PhiSpace(
  reference,
  query,
  phenotypes = NULL,
  response = NULL,
  PhiSpaceAssay = "rank",
  regMethod = c("PLS", "PCA"),
  ncomp = NULL,
  nfeat = NULL,
  selectedFeat = NULL,
  center = TRUE,
  scale = FALSE,
  DRinfo = FALSE,
  storeUnNorm = FALSE,
  updateRef = FALSE
)
}
\arguments{
\item{reference}{SCE. The reference.}

\item{query}{SCE or a list of SCE. The query.}

\item{phenotypes}{Charater. Which types of phenotypes to predict. If \code{NULL}, then have to specify \code{response}.}

\item{response}{Named matrix. Rows correpond to cells (columns) in reference; columns correspond to phenotypes. If not \code{NULL}, then will override \code{phenotypes}.}

\item{PhiSpaceAssay}{Character. Which assay to use to train}

\item{regMethod}{Character. Regression method: one of PLS and PCA}

\item{ncomp}{Integer.}

\item{nfeat}{Integer.}

\item{selectedFeat}{Character.}

\item{center}{Logic.}

\item{scale}{Logic.}

\item{DRinfo}{Logic. Whether to return dimension reduction information from PCA or PLS. Disable to save memory.}

\item{storeUnNorm}{Store unnormalised raw PhiSpace scores or not. Default is \code{FALSE}.}

\item{updateRef}{Update reference (store reference PhiSpace scores in reference sce object) or not.}
}
\value{
An SCE object with annotation results stored as a reducedDims slot.
}
\description{
PhiSpace continuous phenotyping
}
