% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhiSpaceR_1ref.R
\name{PhiSpaceR_1ref}
\alias{PhiSpaceR_1ref}
\title{PhiSpace using a single reference}
\usage{
PhiSpaceR_1ref(
  reference,
  query,
  phenotypes,
  PhiSpaceAssay = "rank",
  regMethod = c("PLS", "PCA"),
  ncomp = NULL,
  nfeat = NULL,
  selectedFeat = NULL,
  center = TRUE,
  scale = FALSE
)
}
\arguments{
\item{reference}{SingleCellExperiment object. The reference}

\item{query}{SingleCellExperiment object. The query}

\item{phenotypes}{Charater. Which types of phenotypes to predict}

\item{PhiSpaceAssay}{Character. Which assay to use as predictors}

\item{regMethod}{Character. Regression method: one of PLS and PCA}

\item{ncomp}{Integer.}

\item{nfeat}{Integer.}

\item{selectedFeat}{Character.}

\item{center}{Logic.}

\item{scale}{Logic.}
}
\value{
A list
}
\description{
PhiSpace using a single reference
}
