% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.fit.R
\name{pls.fit}
\alias{pls.fit}
\title{Title}
\usage{
pls.fit(X, Y, ncomp, center = TRUE, stripped = FALSE)
}
\arguments{
\item{stripped}{}
}
\description{
Same as \code{pls::kernelpls.fit}, but replacing a full eigen decomposition by a partial one via \code{rARPACK::eigs_sym}.
}
