% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.fit.R
\name{pls.fit}
\alias{pls.fit}
\title{Title}
\usage{
pls.fit(X, Y, ncomp, center = TRUE, stripped = FALSE)
}
\arguments{
\item{stripped}{}
}
\value{
A list containing
\item{coefficients}{Regression coefficient matrices.}
\item{scores}{Score matrix for X.}
\item{loadings}{Laoding matrix for X.}
\item{Yloadings}{Loading matrix for Y.}
\item{projection}{}
\item{Xmeans}{}
\item{Ymeans}{}
\item{fitted.values}{}
\item{residuals}{}
\item{Xvar}{}
\item{Xtotvar}{}
}
\description{
Same as \code{pls::kernelpls.fit}, but replacing a full eigen decomposition by a partial one via \code{rARPACK::eigs_sym}.
}
