% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulk.R
\name{pseudoBulk}
\alias{pseudoBulk}
\title{Converting single cells to pseudo-bulks.}
\usage{
pseudoBulk(
  query,
  clusterid,
  assay = "counts",
  size.factor = 1,
  pool.factor = 1,
  nPool = 15,
  YY = NULL
)
}
\arguments{
\item{query}{\code{SingleCellExpeirment} object.}

\item{clusterid}{Character.}

\item{assay}{Character.}

\item{size.factor}{Numeric.}

\item{pool.factor}{Numeric.}

\item{nPool}{Integer.}

\item{YY}{Matrix.}
}
\value{
An updated \code{SingleCellExpeirment} object with a pseudobulk assay
}
\description{
Converting single cells to pseudo-bulks.
}
