% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankTransf.R
\name{RankTransf}
\alias{RankTransf}
\title{Rank transform of a SingleCellExperiment object.}
\usage{
RankTransf(sce, assayname = "counts", targetAssay = "rank", sparse = TRUE)
}
\arguments{
\item{sparse}{}
}
\description{
Rank transform of a SingleCellExperiment object.
}
