% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcTransf.R
\name{rcTransf}
\alias{rcTransf}
\title{Relative counts normalisation.}
\usage{
rcTransf(sce, assayName = "counts", targetAssay = "data")
}
\arguments{
\item{targetAssay}{}
}
\description{
Relative counts normalisation.
}
