% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RTassay}
\alias{RTassay}
\title{Apply rank transform to a gene by cell matrix.}
\usage{
RTassay(X)
}
\arguments{
\item{X}{A cell by gene matrix.}
}
\description{
Apply rank transform to a gene by cell matrix.
}
