% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSampler.R
\name{sample_grid}
\alias{sample_grid}
\title{Grid-based spatial sampling}
\usage{
sample_grid(
  coordinates,
  n_target,
  grid_size,
  min_cells_per_region,
  balance_regions
)
}
\description{
Grid-based spatial sampling
}
