% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSampler.R
\name{sample_kmeans}
\alias{sample_kmeans}
\title{K-means based spatial sampling}
\usage{
sample_kmeans(coordinates, n_target, min_cells_per_region, balance_regions)
}
\description{
K-means based spatial sampling
}
