% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPC.R
\name{SuperPC}
\alias{SuperPC}
\title{Compute regression model.}
\usage{
SuperPC(
  re = NULL,
  reference = NULL,
  YY = NULL,
  labelName = NULL,
  ncomp = NULL,
  Nselect = NULL,
  impScores = NULL,
  selectFeat = NULL,
  assayName = "logcounts",
  regMethod = "PCA",
  labelCode = "-1,1"
)
}
\arguments{
\item{labelCode}{}
}
\description{
Compute regression model.
}
