% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdspc.fit.R
\name{svdspc.fit}
\alias{svdspc.fit}
\title{Principle component regression}
\usage{
svdspc.fit(X, Y, ncomp, center = TRUE, stripped = FALSE)
}
\arguments{
\item{stripped}{}
}
\value{
A list containing
\item{coefficients}{Regression coefficient matrices.}
\item{scores}{Score matrix for X.}
\item{loadings}{Laoding matrix for X.}
\item{Yloadings}{Loading matrix for Y.}
\item{projection}{}
\item{Xmeans}{}
\item{Ymeans}{}
\item{fitted.values}{}
\item{residuals}{}
\item{Xvar}{}
\item{Xtotvar}{}
}
\description{
Same as \code{pls::svdpc.fit}, except using \code{rARPACK::svds} instead of \code{La.svd}.
}
