% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdspc.fit.R
\name{svdspc.fit}
\alias{svdspc.fit}
\title{Principle component regression}
\usage{
svdspc.fit(X, Y, ncomp)
}
\arguments{
\item{X}{Matrix. Scaled predictor matrix}

\item{Y}{Matrix. Scaled response matrix}

\item{ncomp}{Integer.}
}
\value{
A list containing
\item{coefficients}{Regression coefficient matrices.}
}
\description{
Same as \code{pls::svdpc.fit}, except using \code{rARPACK::svds} instead of \code{La.svd}.
}
