% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroFeatQC.R
\name{zeroFeatQC}
\alias{zeroFeatQC}
\title{Minimal quality control: remove features with all zero values.}
\usage{
zeroFeatQC(sce, assayName = "counts")
}
\arguments{
\item{assayName}{}
}
\description{
Minimal quality control: remove features with all zero values.
}
